2DCNN/requirements_conda.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
blas=1.0=mkl
ca-certificates=2021.4.13=haa95532_1
certifi=2020.12.5=py36haa95532_0
chardet=4.0.0=pypi_0
cudatoolkit=10.1.168=0
cycler=0.10.0=py36haa95532_0
decorator=4.4.2=pypi_0
et_xmlfile=1.0.1=py_1001
freetype=2.10.4=hd328e21_0
h5py=2.8.0=py36h3bdd7fb_2
hdf5=1.10.2=hac2f561_1
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha925a31_3
idna=2.10=pypi_0
imageio=2.9.0=pyhd3eb1b0_0
intel-openmp=2020.2=254
jdcal=1.4.1=py_0
joblib=1.0.1=pypi_0
jpeg=9b=hb83a4c4_2
jsonpatch=1.32=pypi_0
jsonpointer=2.1=pypi_0
kiwisolver=1.3.1=py36hd77b12b_0
libpng=1.6.37=h2a8f88b_0
libtiff=4.1.0=h56a325e_1
lz4-c=1.9.3=h2bbff1b_0
matplotlib=3.3.4=py36haa95532_0
matplotlib-base=3.3.4=py36h49ac443_0
mkl=2020.2=256
mkl-service=2.3.0=py36h196d8e1_0
mkl_fft=1.3.0=py36h46781fe_0
mkl_random=1.1.1=py36h47e9c7a_0
networkx=2.5.1=pypi_0
ninja=1.10.2=h6d14046_1
numpy=1.19.5=pypi_0
numpy-base=1.19.2=py36ha3acd2a_0
olefile=0.46=py36_0
opencv-contrib-python=4.5.1.48=pypi_0
opencv-python=3.4.2.16=pypi_0
openpyxl=3.0.7=pyhd3eb1b0_0
openssl=1.1.1k=h2bbff1b_0
pandas=1.1.3=py36ha925a31_0
pillow=8.2.0=py36h4fa10fc_0
pip=21.0.1=py36haa95532_0
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py36h6538335_2
python=3.6.2=h09676a0_15
python-dateutil=2.8.1=pyhd3eb1b0_0
pytorch=1.5.0=py3.6_cuda101_cudnn7_0
pytz=2021.1=pyhd3eb1b0_0
pywavelets=1.1.1=pypi_0
pyzmq=22.0.3=pypi_0
qt=5.9.7=vc14h73c81de_0
requests=2.25.1=pypi_0
scikit-image=0.17.2=pypi_0
scikit-learn=0.24.1=pypi_0
scipy=1.5.4=pypi_0
seaborn=0.11.1=pyhd3eb1b0_0
setuptools=52.0.0=py36haa95532_0
sip=4.19.8=py36h6538335_0
six=1.15.0=py36haa95532_0
spectral=0.19=pypi_0
sqlite=3.35.4=h2bbff1b_0
threadpoolctl=2.1.0=pypi_0
tifffile=2020.9.3=pypi_0
tk=8.6.10=he774522_0
torchfile=0.1.0=pypi_0
torchsummary=1.5.1=pypi_0
torchvision=0.6.0=py36_cu101
tornado=6.1=py36h2bbff1b_0
tqdm=4.59.0=pyhd3eb1b0_1
urllib3=1.26.4=pypi_0
vc=14.2=h21ff451_1
visdom=0.1.8.9=pypi_0
vs2015_runtime=14.27.29016=h5e58377_2
websocket-client=0.58.0=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
wincertstore=0.2=py36h7fe50ca_0
xz=5.2.5=h62dcd97_0
zlib=1.2.11=h62dcd97_4
zstd=1.4.9=h19a0ad4_0
